requirements.txt
argon2-cffi==23.1.0
argon2-cffi-bindings==21.2.0
asgiref==3.8.1
cffi==1.17.1
crispy-bootstrap5==2024.10
Django==5.1.2
django-allauth==65.0.2
django-browser-reload==1.16.0
django-crispy-forms==2.3
django-debug-toolbar==4.4.6
django-environ==0.11.2
django-extensions==3.2.3
django-simple-history==3.7.0
django-tenants==3.7.0
environ==1.0
psycopg2==2.9.10
psycopg2-binary==2.9.10
pycparser==2.22
setuptools==75.1.0
sqlparse==0.5.1
wheel==0.44.0
